requirements.txt
langchain>=0.1.0
openai>=1.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.3
python-dotenv>=1.0.0
streamlit>=1.30.0
tiktoken>=0.5.2
numpy>=1.24.0
pandas>=2.0.0
python-magic>=0.4.27
pypdf>=3.17.0
sentence-transformers>=2.2.2
Pillow>=10.0.0
pdf2image>=1.16.3
python-docx>=0.8.11
opencv-python>=4.8.0
pytesseract>=0.3.10 